% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_c_payment.R
\name{calculate_cc_payment}
\alias{calculate_cc_payment}
\title{Calculate CC Payment}
\usage{
calculate_cc_payment(
  balance,
  first_due_date,
  last_due_date,
  credit_limit,
  ending_balance
)
}
\description{
Calculate CC Payment
}
