% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apr_calculator.R
\name{calculate_multiplier}
\alias{calculate_multiplier}
\title{Calculate Multiplier}
\usage{
calculate_multiplier(
  apr = 0.5,
  compound_period = "month",
  n_compound_period = 12
)
}
\description{
Calculate Multiplier
}
