% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_c_payment.R
\name{calculate_usbank_payment}
\alias{calculate_usbank_payment}
\title{Calculate US Bank Payment}
\usage{
calculate_usbank_payment(
  balance,
  first_due_date = "2021-09-16",
  last_due_date = "2023-04-16",
  credit_limit = 10000,
  ending_balance = 3000
)
}
\description{
Calculate US Bank Payment
}
