% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_c_payment.R
\name{calculate_wf_payment}
\alias{calculate_wf_payment}
\title{Calculate WF Payment}
\usage{
calculate_wf_payment(
  balance,
  first_due_date = "2021-12-21",
  last_due_date = "2023-08-21",
  credit_limit = 18000,
  ending_balance = 5400
)
}
\description{
Calculate WF Payment
}
