% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculus-package.R
\docType{package}
\name{calculus-package}
\alias{calculus}
\alias{calculus-package}
\title{calculus: Various financial and retail calculators.}
\description{
Analyze savings and purchases.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/meerapatelmd/calculus}
  \item Report bugs at \url{https://github.com/meerapatelmd/calculus/issues}
}

}
\author{
\strong{Maintainer}: Meera Patel, MD \email{patelmeeray@gmail.com} (\href{https://orcid.org/0000-0002-3401-3948}{ORCID})

}
\keyword{internal}
