% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apr_calculator.R
\name{estimate_account_balance}
\alias{estimate_account_balance}
\title{Estimate Account Balance}
\usage{
estimate_account_balance(
  principle = 30000,
  apr = 0.5,
  compound_period = "month",
  deposit_per_compound_period = 0,
  n_compound_period = 12
)
}
\description{
Estimate Account Balance
}
