% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20211224_market_value.R
\name{evaluate_stud_earrings}
\alias{evaluate_stud_earrings}
\title{Evaluate Diamond Stud Earrings}
\usage{
evaluate_stud_earrings(
  listed_cost,
  total_carats,
  color,
  clarity,
  metal,
  gia_cut_grade = "",
  fluorescence = ""
)
}
\description{
Assumes brilliant cut diamond.
}
\details{
https://www.diamonds.pro/education/diamond-prices/
}
\seealso{
Other Market Value: 
\code{\link{value_gold}()}
}
\concept{Market Value}
