% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunset.R
\name{get_daylight_times}
\alias{get_daylight_times}
\title{Sunrise/Sunset Data}
\usage{
get_daylight_times(
  latitude = 40.73,
  longitude = -74,
  start_date = "2022-01-10",
  end_date = "2022-12-31"
)
}
\description{
Get the parsed response from https://api.sunrise-sunset.org
based on the given date range.
}
