% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_calculator.R
\name{next_business_days}
\alias{next_business_days}
\title{Next Business Days of a Vector of Dates}
\usage{
next_business_days(input_dates, verbose = TRUE)
}
\description{
Next Business Days of a Vector of Dates
}
