% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_calculator.R
\name{previous_business_day}
\alias{previous_business_day}
\title{Calculate the Previous Business Day}
\usage{
previous_business_day(input_date, verbose = TRUE)
}
\description{
Calculate the Previous Business Day
}
