% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_calculator.R
\name{previous_business_days}
\alias{previous_business_days}
\title{Previous Business Days of a Vector of Dates}
\usage{
previous_business_days(input_dates, verbose = TRUE)
}
\description{
Previous Business Days of a Vector of Dates
}
