% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devel-market_value.R
\name{value_gold}
\alias{value_gold}
\title{Value Gold Jewelry}
\usage{
value_gold(usd_per_ounce, ounces = 0.1, karats = c("24K", "18K", "14K", "10K"))
}
\description{
Value Gold Jewelry
}
\seealso{
Other Market Value: 
\code{\link{evaluate_stud_earrings}()}
}
\concept{Market Value}
