% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit_card_payment_calculator.R
\name{wf_reflect_payment}
\alias{wf_reflect_payment}
\title{Wells Fargo Reflect Credit Card Payment}
\usage{
wf_reflect_payment(balance)
}
\description{
Wells Fargo Reflect Credit Card Payment
}
