% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gcal_csv.R
\name{write_gcal_csv}
\alias{write_gcal_csv}
\title{Write CSV to Import Into Google Calendar}
\usage{
write_gcal_csv(
  subject = "New Event",
  start_date_day_number,
  start_date_year,
  start_date_offset = c("none", "previous_business_day", "next_business_day"),
  gcal_csv,
  verbose = TRUE
)
}
\description{
Write CSV to Import Into Google Calendar
}
